fix: 딜러 통계 화면 개선 및 consult dealer_no 타입 오류 수정

+++ src/main/java/com/kiaev/client/consult/ConsultService.java
@@ -37,7 +37,7 @@ public class ConsultService {
     }
 
     // 상담 상태 변경
-    public void updateConsultStatus(Long consultNo, String consultStatus, Long dealerNo) {
+    public void updateConsultStatus(Long consultNo, String consultStatus, Integer dealerNo) {
         Consult consult = consultationRepository.findById(consultNo).orElse(null);
 
         if (consult == null) {
